+++ plumbum/utils.py
@@ -1,15 +1,7 @@
 from __future__ import with_statement
 import six
-import sys
-import os
-import subprocess
-import signal
-import errno
-import time
-import traceback
 from plumbum.path import Path
 from plumbum.local_machine import local, LocalPath
-from plumbum.commands import ProcessExecutionError
 
 
 def delete(*paths):
@@ -37,12 +29,23 @@ def move(src, dst):
     """Moves the source path onto the destination path; ``src`` and ``dst`` can be either
     strings, :class:`LocalPaths <plumbum.local_machine.LocalPath>` or
     :class:`RemotePath <plumbum.remote_machine.RemotePath>`; any combination of the three will
-    work.
+    work. 
+    
+    .. versionadded:: 1.3
+        ``src`` can also be a list of strings/paths, in which case ``dst`` must not exist or be a directory.
     """
-    if not isinstance(src, Path):
-        src = local.path(src)
     if not isinstance(dst, Path):
         dst = local.path(dst)
+    if isinstance(src, (tuple, list)):
+        if not dst.exists():
+            dst.mkdir()
+        elif not dst.isdir():
+            raise ValueError("When using multiple sources, dst %r must be a directory" % (dst,))
+        for src2 in src:
+            move(src2, dst)
+        return
+    elif not isinstance(src, Path):
+        src = local.path(src)
 
     if isinstance(src, LocalPath):
         if isinstance(dst, LocalPath):
@@ -63,11 +66,22 @@ def copy(src, dst):
     either strings, :class:`LocalPaths <plumbum.local_machine.LocalPath>` or
     :class:`RemotePath <plumbum.remote_machine.RemotePath>`; any combination of the three will
     work.
+
+    .. versionadded:: 1.3
+        ``src`` can also be a list of strings/paths, in which case ``dst`` must not exist or be a directory.
     """
-    if not isinstance(src, Path):
-        src = local.path(src)
     if not isinstance(dst, Path):
         dst = local.path(dst)
+    if isinstance(src, (tuple, list)):
+        if not dst.exists():
+            dst.mkdir()
+        elif not dst.isdir():
+            raise ValueError("When using multiple sources, dst %r must be a directory" % (dst,))
+        for src2 in src:
+            copy(src2, dst)
+        return
+    elif not isinstance(src, Path):
+        src = local.path(src)
 
     if isinstance(src, LocalPath):
         if isinstance(dst, LocalPath):
@@ -86,99 +100,5 @@ def copy(src, dst):
                 copy(src, tmp)
                 copy(tmp / src.basename, dst)
             return dst
-
-def _posix_daemonize(command, cwd = "/"):
-    """run ``command`` as a daemon: fork a child process to setpid, redirect std handles to /dev/null, 
-    umask, close all fds, chdir to ``cwd``, then fork and exec ``command``. Returns a ``Popen`` process that
-    can be used to poll/wait for the executed command (but keep in mind that you cannot access std handles) 
-    """
-    MAX_SIZE = 16384
-    rfd, wfd = os.pipe()
-    argv = command.formulate()
-    firstpid = os.fork()
-    if firstpid == 0:
-        # first child: become session leader,
-        os.close(rfd)
-        rc = 0
-        try:
-            os.setsid()
-            os.umask(0)
-            stdin = open(os.devnull, "r")
-            stdout = open(os.devnull, "w")
-            stderr = open(os.devnull, "w")
-            signal.signal(signal.SIGHUP, signal.SIG_IGN)
-            proc = command.popen(cwd = cwd, close_fds = True, stdin = stdin.fileno(), 
-                stdout = stdout.fileno(), stderr = stderr.fileno())
-            os.write(wfd, str(proc.pid))
-        except:
-            rc = 1
-            tbtext = "".join(traceback.format_exception(*sys.exc_info()))[-MAX_SIZE:]
-            os.write(wfd, tbtext)
-        finally:
-            os.close(wfd)
-            os._exit(rc)
-    else:
-        # wait for first child to die
-        os.close(wfd)
-        _, rc = os.waitpid(firstpid, 0)
-        output = os.read(rfd, MAX_SIZE)
-        if rc == 0 and output.isdigit():
-            secondpid = int(output)
-        else:
-            raise ProcessExecutionError(argv, rc, "", output)
-        proc = subprocess.Popen.__new__(subprocess.Popen)
-        proc._child_created = True
-        proc.returncode = None
-        proc.stdout = None
-        proc.stdin = None
-        proc.stderr = None
-        proc.pid = secondpid
-        proc.universal_newlines = False
-        proc._input = None
-        proc._communication_started = False
-        proc.args = argv
-        proc.argv = argv
-        
-        def poll(self = proc):
-            if self.returncode is None:
-                try:
-                    os.kill(self.pid, 0)
-                except OSError:
-                    ex = sys.exc_info()[1]
-                    if ex.errno == errno.ESRCH:
-                        # process does not exist
-                        self.returncode = 0
-                    else:
-                        raise
-            return self.returncode
-        
-        def wait(self = proc):
-            while self.returncode is None:
-                if self.poll() is None:
-                    time.sleep(0.5)
-            return proc.returncode                
-        
-        proc.poll = poll
-        proc.wait = wait
-        return proc
-
-
-def _win32_daemonize(command, cwd = "/"):
-    """run ``command`` as a "windows daemon": detach from controlling console and create a new process group.
-    This means that the command will not receive console events and would survive its parent's termination. 
-    Returns a ``Popen`` object.
-    """
-    DETACHED_PROCESS = 0x00000008
-    stdin = open(os.devnull, "r")
-    stdout = open(os.devnull, "w")
-    stderr = open(os.devnull, "w")
-    return command.popen(cwd = cwd, stdin = stdin.fileno(), stdout = stdout.fileno(), stderr = stderr.fileno(), 
-        creationflags = subprocess.CREATE_NEW_PROCESS_GROUP | DETACHED_PROCESS)
-
-
-if sys.platform == "win32":
-    daemonize = _win32_daemonize
-else:
-    daemonize = _posix_daemonize